+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to [Sikessem's UI kit](https://github.com/sikessem/ui) will 
 
 **Full Changelog:** [v0.5.1...v0.5.2](https://github.com/sikessem/ui/compare/v0.5.1...v0.5.2)
 
+- [0.x] Fix [icon component](https://github.com/sikessem/ui#icon-component) fill color
+
 ## [v0.5.2](https://github.com/sikessem/ui/releases/tag/v0.5.2) - 2023-08-03
 
 - [0.x] Fix component slot rendering